requirements.txt
Babel==2.9.1
Deprecated==1.2.13
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
wrapt==1.14.0
